pyproject.toml
[tool.poetry]
name = "oops"
version = "0.1.0"
description = "Open Source Offensive Security Agent"
authors = ["Farhan Aulianda <farhanauliandaa@gmail.com>"]
readme = "README.md"
packages = [{include = "oops"}]

[tool.poetry.scripts]
oops = "oops.main:main"

[tool.poetry.dependencies]
python = "^3.10"
strands-agents = {extras = ["openai"], version = "^1.19.0"}
strands-agents-tools = "^0.2.17"
python-dotenv = "^1.2.1"
dnspython = "^2.8.0"
requests = "^2.32.5"
checkdmarc = "^5.12.27"
rich = "^14.2.0"
docker = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
